refactor: rename API validator files to avoid path conflicts
