+++ scripts/README.md
@@ -10,12 +10,12 @@ scripts
 ├───tests  # contains all unit tests from the validate package
 │       test_validate_format.py
 │       test_validate_links.py
-│       test_validate_apis.py
+│       test_validate_api_health.py
 │
 └───validate  # validate package
         format.py
         links.py
-        apis.py
+        api_health.py
 ```
 
 ## Install dependencies
@@ -53,13 +53,13 @@ $ python scripts/validate/links.py README.md -odlc
 To run API health and free-tier signal validation on the `README.md` file, run:
 
 ```bash
-$ python scripts/validate/apis.py README.md
+$ python scripts/validate/api_health.py README.md
 ```
 
 To apply automatic removals for rows that fail the checks, run:
 
 ```bash
-$ python scripts/validate/apis.py README.md --write
+$ python scripts/validate/api_health.py README.md --write
 ```
 
 ## Running Tests
@@ -91,5 +91,5 @@ $ python -m unittest discover tests/ --verbose --pattern "test_validate_links.py
 To run only the API validation tests, run:
 
 ```bash
-$ python -m unittest discover tests/ --verbose --pattern "test_validate_apis.py"
+$ python -m unittest discover tests/ --verbose --pattern "test_validate_api_health.py"
 ```
